Lock Swift version to focal.
The current `latest` build of the Swift image uses a compiler version of Swift 5.6.x.
Unfortunately, but actually rather strange, there is no current arm64 image for swift… Instead of building our own Swift image as well, we use the focal tag, as this image uses Swift v5.5 which is available as a native arm64 image.