Add the magic matrix script.
This bash script generates the build matrix for only the changed images. It will only append those directories that contain a changed (or added) `Dockerfile` or `RELEASE`-file. It will then extract the version from the `RELEASE`-file and adds it to the matrix – this information will be used by the buildx-step which injects it into the Dockerfile.

Finally, it generates two flags `build_base_images` and `build_bamboo_variants` that will also be used by the workflow to skip steps completely.